in utils.js   /**    * Return the height/width of the visible portion of the supplied window.    * @param Window.    * @return [width, height]    */   contentWidthHeight: function(win) { 		if(self.innerWidth != undefined) { 		  return [win.innerWidth, win.innerHeight];     } else {   		//var docelem = win.document.documentElement;   		var docelem = win.document.body;   		return [docelem.clientWidth, docelem.clientHeight]; 		}   },
Changed this line because of http://www.howtocreate.co.uk/tutorials/javascript/browserwindow top.document.clientWidth is undefined, top.document.documentElement.clientWidth is 0, top.document.body.clientWith is 1036 (at least during one test).. This could either refer to document width (in quirksmode) or window width (in standard mode). Either way...it will never return 0...

This fixed the whole interface from constantly shifting to the left but the problem remains that as soon as you press down arrow, there is a slight shift to the left. BUT as soon as you press control+l to go to a new location it shifts back to the starting point (this is where it would keep shifting left before). 

I think the shift to the left might be in the getScrollOffset function (in utils.js). I have a feeling document.documentElement.scrollLeft might have a similar issue to clientWidth. Searching.... 

From that same article, there does seem to be similar issues with IE and scroll x/y values. Therefore, I added the document.body.scrollLeft bit to the following code in utils.js

  getScrollOffset: function(win) {
  	if(win.scrollY != undefined) {
  		return [win.scrollX, win.scrollY];
  	} else if(win.document.body.scrollTop) {
  	   var delem = win.document.body;
  	   return [delem.scrollLeft, delem.scrollTop];
  	}
  	else if(win.document.documentElement.scrollTop) {
  		var delem = win.document.documentElement;
      return [delem.scrollLeft, delem.scrollTop];
  	} else {
  		return [0, 0];
  	}
  }

But this did not seem to affect the behavior related to the slight shift left when pressing the down arrow key.

Is it related to positioning the content window? When it grabs the dimensions and gets window instead of document...where can I see those? how do I know which value is being returned? 

Moving onto the "why doesn't currently read text display in the "wa_text_display" span in IE...I commented out the following from interface.js

updateCurrentlyPlayingSize: function(time, total) {
    // Not the best way to handle whatever problem is happening with IE.
    //if(WA.Utils.isIE()) {
     // return;
    //}
Right...so now every now and again the content scrolls by...so it seems that the values for IE throw it into the if condition that does scrolling...

Right, in interface.js in updateCurrentlyPlayingSize, width=1036, text_disp = 5000 (since that is set in index.php on the wa_text_display_container), so that diff is -3964. Thus the text is scrolling because the only choices in the if/else are for values >=0 (where you don't scroll) and everything else for which you do scroll.

So, I think Jeffrey was expecting that the offset width of that span is the width of the characters but it is always returning 5000 because that is the length of the div and the span is set to 100% width...

What if I comment out the following in index.php (got rid of width=5000px on the container div)
 <!-- div id="wa_text_display_container" style="width: 5000px;" -->
        <div id="wa_text_display_container">
          <span id="wa_text_display" style="width: 100%; margin: 0; padding: 0.1em; font-size: 3em; color: #FF0; font-weight: bold;"><?php echo wa_gettext("Welcome to WebAnywhere") ?></span>
        </div>

Still seems to work on FF/Mac. And it works in IE/Win!!!!!